move bytearrstream to own file
