+++ src/net/server/handlers/mod.rs
@@ -1,7 +1,5 @@
-mod create_query;
 mod echo;
-mod process_query;
+mod query;
 
-pub use create_query::handler as create_query_handler;
 pub use echo::{handler as echo_handler, Payload as EchoData};
-pub use process_query::{handler as process_query_handler, obtain_permit_mw};
+pub use query::{handler as process_query_handler, obtain_permit_mw};
